+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.12] - 2025-11-05
+
+### Added
+- [#126] Add --only-statuses flag to get_requests.pl to show status counts
+
 ## [1.0.11] - 2025-11-05
 
 ### Fixed
@@ -13,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - [#126] Add --lastcircstate parameter to get_requests.pl for filtering by last circulation state
+- [#126] Add --only-statuses flag to get_requests.pl to show status counts
 
 ## [1.0.10] - 2025-11-04